Fix integer underflow in saved_bytes calculation
  Cast src and dst to i64 before subtraction to avoid u64 underflow
  when the compressed file is larger than the source.

+++ src/precompress.rs
@@ -236,7 +236,7 @@ impl Compressor {
                         Algorithm::Zstd => &mut stats.zstd,
                     };
                     s.total_time += dur;
-                    s.saved_bytes += (src - dst) as i64;
+                    s.saved_bytes += src as i64 - dst as i64;
                     s.total_bytes += dst;
                     stats.num_files += 1;
                 }
